fix(test): user select field test fail issue
